requirements.txt
Flask==2.3.3
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
fpdf2==2.7.5
python-dotenv==1.0.0
polyline==2.0.0
matplotlib==3.9.4
contourpy==1.3.0
cycler==0.12.1
kiwisolver==1.4.7
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0
xlsxwriter==3.1.6
unicodedata2==15.1.0
geopy==2.4.1
folium==0.15.1
branca==0.7.0
packaging==23.2
pillow==10.2.0
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.5
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
tzdata==2024.1
Flask-Caching==2.0.2
cachelib==0.9.0
brotli==1.1.0
redis==5.0.3
reportlab
psycopg2-binary
flask-sqlalchemy
flask-migrate 